Simplify kp Bandit Trio achievement checking
Previously, we have checked if any of the two achievements relevant
to the encounter is completed because we weren't sure there is an
achievement that is awarded just for completing the encounter.

It appears the achievement (The Shield) just has a bit confusing
description (you can't really lose any civilians here without failing
completely).